requirement.txt
streamlit
python-dotenv
openai
google-generativeai